Feature/time ops api (#135)
* factor out time api

* implement TimeOps for Vertices and iterators

* implement timeops for everything

* fix issues after rebase

* clean up warnings and fix property api issues

* add missing timeops implementations and fix semantics for nested window

* use traits to ensure api is consistent and we don't miss methods

* start updating python interface

* fix window semantics and update tests

+++ docbrown/db/src/algorithms/local_clustering_coefficient.rs
@@ -25,7 +25,7 @@
 //! ```rust
 //! use docbrown_db::algorithms::local_clustering_coefficient::{local_clustering_coefficient};
 //! use docbrown_db::graph::Graph;
-//! use docbrown_db::view_api::GraphViewOps;
+//! use docbrown_db::view_api::*;
 //!
 //! let g = Graph::new(1);
 //! let windowed_graph = g.window(0, 7);
@@ -51,7 +51,6 @@
 
 use crate::algorithms::local_triangle_count::local_triangle_count;
 use crate::view_api::*;
-use docbrown_core::tgraph_shard::errors::GraphError;
 
 /// measures the degree to which nodes in a graph tend to cluster together
 pub fn local_clustering_coefficient<G: GraphViewOps>(graph: &G, v: u64) -> Option<f32> {
@@ -60,7 +59,7 @@ pub fn local_clustering_coefficient<G: GraphViewOps>(graph: &G, v: u64) -> Optio
             let triangle_count = triangle_count as f32;
             let degree = vertex.degree() as f32;
             if degree > 1.0 {
-                Some((2.0 * triangle_count) / (degree * (degree - 1.0) as f32))
+                Some((2.0 * triangle_count) / (degree * (degree - 1.0)))
             } else {
                 Some(0.0)
             }
